+++ packages/ffe-file-upload-react/src/file-content.js
@@ -1,0 +1,8 @@
+export const getFileContent = file => {
+    return new Promise(function(resolve, reject) {
+        const reader = new window.FileReader();
+        reader.onload = event => resolve(event.target.result);
+        reader.onerror = error => reject(error);
+        reader.readAsDataURL(file);
+    });
+};