+++ src/components/link/index.js
@@ -1,0 +1,3 @@
+import LinkRenderer from './LinkRenderer';
+
+export default LinkRenderer;