+++ src/components/style-guide/index.js
@@ -1,0 +1,3 @@
+import StyleGuide from './StyleGuide';
+
+export default StyleGuide;