+++ packages/ffe-account-selector-react/CHANGELOG.md
@@ -3,13 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-      <a name="8.0.4"></a>
+      <a name="8.1.2"></a>
+
+## [8.1.2](***REMOVED***) (2018-01-17)
+
+### Bug Fixes
+
+* **ffe-account-selector-react:** fix build ([f0cedda](***REMOVED***))
+
+
+      <a name="8.1.1"></a>
+
+## [8.1.1](***REMOVED***) (2018-01-16)
+
+**Note:** Version bump only for package ffe-account-selector-react
+
+<a name="8.1.0"></a>
+
+# [8.1.0](***REMOVED***) (2018-01-15)
+
+### Features
+
+* **ffe-account-selector-react:** added prop to disable input-field ([9f38d25](***REMOVED***))
+
+<a name="8.0.4"></a>
 
 ## [8.0.4](***REMOVED***) (2018-01-10)
 
 **Note:** Version bump only for package ffe-account-selector-react
 
-      <a name="8.0.3"></a>
+<a name="8.0.3"></a>
 
 ## [8.0.3](***REMOVED***) (2018-01-10)