+++ packages/ffe-decorators-react/CHANGELOG.md
@@ -3,11 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.4"></a>
+
+## [1.0.4](***REMOVED***) (2018-01-12)
+
+**Note:** Version bump only for package ffe-decorators-react
+
 <a name="1.0.3"></a>
+
 ## [1.0.3](***REMOVED***) (2018-01-12)
-
-
-
 
 **Note:** Version bump only for package ffe-decorators-react