+++ packages/ffe-details-list/CHANGELOG.md
@@ -3,11 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.1.3"></a>
+
+## [8.1.3](***REMOVED***) (2018-01-12)
+
+**Note:** Version bump only for package ffe-details-list
+
 <a name="8.1.2"></a>
+
 ## [8.1.2](***REMOVED***) (2018-01-12)
-
-
-
 
 **Note:** Version bump only for package ffe-details-list